fix: change entitie dataype for id and timestamp attributes

+++ ucm-discuss-be/src/main/java/com/example/ucm_discuss_be/courses/CourseModel.java
@@ -1,6 +1,6 @@
 package com.example.ucm_discuss_be.courses;
 
-import java.math.BigInteger;
+// import java.math.BigInteger;
 // import java.security.Timestamp;
 // import java.sql.Time;
 // import java.util.Objects;
@@ -22,7 +22,7 @@ import jakarta.persistence.Id;
 @Entity
 public class CourseModel {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
     private String course_code;
     private String course_name;
     private String year;
