+++ ucm-discuss-be/src/main/java/com/example/ucm_discuss_be/majors/MajorModel.java
@@ -1,6 +1,6 @@
 package com.example.ucm_discuss_be.majors;
 
-import java.math.BigInteger;
+// import java.math.BigInteger;
 // import java.security.Timestamp;
 // import java.sql.Time;
 // import java.util.Objects;
@@ -22,6 +22,6 @@ import jakarta.persistence.Id;
 
 public class MajorModel {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
     private String name;
 }
