+++ ucm-discuss-be/src/main/java/com/example/ucm_discuss_be/replies/ReplyModel.java
@@ -1,5 +1,5 @@
 package com.example.ucm_discuss_be.replies;
-import java.math.BigInteger;
+// import java.math.BigInteger;
 // import java.security.Timestamp;
 // import java.sql.Time;
 // import java.util.Objects;
@@ -21,7 +21,7 @@ import jakarta.persistence.Id;
 
 public class ReplyModel {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
-    private BigInteger source_comment_id; //Foreign key to comments table
-    private BigInteger parent_comment_id; //Foreign key to comments table
+    private Long id;
+    private Long source_comment_id; //Foreign key to comments table
+    private Long parent_comment_id; //Foreign key to comments table
 }
